Major update of the Arrowhead DSL
Update of the MEta model, profile and libraries and the Papyrus plugin.  The ambition is to reflect the the upcoming Eclipse Arrowhead v5.0 architecture and infrastructure inclduing details on control plane interaction and behaviours. Obviously not fully ready for a wider usage. Please try out and report problems as isses in github.